Finishing the Requirements Outline
I managed to finish the other half of the Requirements Outline and decided how I would collect the green and blue blocks. I will create a cage-like box that only covers the top, and 3 sides so that it is easier to drop it off in the "box" to get the other block.