Trying to get the colour sensor to work
Although we tried by using youtube tutorials, different variables and functions it wont detect the colour but instead detects it as neither green or blue. Yellow and red don't work either. I added the failed attempts into the test cases and I think I will have pages and pages for the test cases soon. I need to get the answer before the holidays so I can code it at home and test it out on the robot at school.